pyproject.toml
[tool.poetry]
name = "guess-ethusdt"
version = "0.1.0"
description = ""
authors = ["Ruslan <shaggybackend@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^1.5.3"
python-binance = "^1.0.17"


[tool.poetry.group.dev.dependencies]
ipython = "^8.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
track = "src.main:main"
